third_party/blink/web_tests/fast/dom/shadow/content-pseudo-element-overridden-expected.txt

CONSOLE WARNING: line 33: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
content

Test for crbug.com/274059. Should be able to override ::content styles in shadow root style sheet from the document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getComputedStyle(document.querySelector(".contentClass")).color is "rgb(0, 0, 255)"
PASS successfullyParsed is true

TEST COMPLETE
